server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18
anthropic>=0.42.0
openai>=1.60.0
google-genai>=1.0.0
pdfplumber>=0.11.0
httpx>=0.28.0
python-dotenv>=1.0.0
tenacity>=9.0.0
motor>=3.6.0
pymongo>=4.6.0
Pillow>=10.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0

